Change pageNums from uint16_t to uint32_t
Users of FPDF_ImportPages API may import more than 65535 pages.  Hence
changing pageNums definition to uint32_t.

+++ fpdfsdk/fpdfppo.cpp
@@ -93,13 +93,13 @@ bool ParserPageRangeString(ByteString rangstring,
     cbMidRange = rangstring.Mid(nStringFrom, nStringTo.value() - nStringFrom);
     auto nMid = cbMidRange.Find('-');
     if (!nMid.has_value()) {
-      uint16_t pageNum =
-          pdfium::base::checked_cast<uint16_t>(atoi(cbMidRange.c_str()));
+      uint32_t pageNum =
+          pdfium::base::checked_cast<uint32_t>(atoi(cbMidRange.c_str()));
       if (pageNum <= 0 || pageNum > nCount)
         return false;
       pageArray->push_back(pageNum);
     } else {
-      uint16_t nStartPageNum = pdfium::base::checked_cast<uint16_t>(
+      uint32_t nStartPageNum = pdfium::base::checked_cast<uint32_t>(
           atoi(cbMidRange.Left(nMid.value()).c_str()));
       if (nStartPageNum == 0)
         return false;
@@ -109,13 +109,13 @@ bool ParserPageRangeString(ByteString rangstring,
       if (nEnd == 0)
         return false;
 
-      uint16_t nEndPageNum = pdfium::base::checked_cast<uint16_t>(
+      uint32_t nEndPageNum = pdfium::base::checked_cast<uint32_t>(
           atoi(cbMidRange.Mid(nMid.value(), nEnd).c_str()));
       if (nStartPageNum < 0 || nStartPageNum > nEndPageNum ||
           nEndPageNum > nCount) {
         return false;
       }
-      for (uint16_t i = nStartPageNum; i <= nEndPageNum; ++i) {
+      for (uint32_t i = nStartPageNum; i <= nEndPageNum; ++i) {
         pageArray->push_back(i);
       }
     }
@@ -147,7 +147,7 @@ class CPDF_PageOrganizer {
   ~CPDF_PageOrganizer();
 
   bool PDFDocInit();
-  bool ExportPage(const std::vector<uint16_t>& pageNums, int nIndex);
+  bool ExportPage(const std::vector<uint32_t>& pageNums, int nIndex);
 
  private:
   using ObjectNumberMap = std::map<uint32_t, uint32_t>;
@@ -206,7 +206,7 @@ bool CPDF_PageOrganizer::PDFDocInit() {
   return true;
 }
 
-bool CPDF_PageOrganizer::ExportPage(const std::vector<uint16_t>& pageNums,
+bool CPDF_PageOrganizer::ExportPage(const std::vector<uint32_t>& pageNums,
                                     int nIndex) {
   int curpage = nIndex;
   auto pObjNumberMap = pdfium::MakeUnique<ObjectNumberMap>();
